+++ src/main/java/com/pradipta/todo/task/TaskService.java
@@ -32,7 +32,7 @@ public class TaskService {
     public Task markDone(int id) {
         //repo.findById(id).get().setIsDone(true);
         //TODO: figure out update in one line lambda
-        Task task = repo.findById((long)id).get();    //TODO: handle empty optional
+        Task task = repo.findById(id).get();    //TODO: handle empty optional
         task.setIsDone(true);
         return repo.save(task);
     }
